requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
pytest
requests
jsonpath